backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
sentence-transformers
transformers
keybert
openai
pydantic
